requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
frozenlist==1.3.1
gino==1.0.1
idna==3.4
multidict==6.0.2
numpy==1.23.3
opencv-python==4.6.0.66
Pillow==9.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qq-botpy==1.1.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.3.24
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
yarl==1.8.1
